Resized the question text box
Only done because the questions aren't likely to be as long as the placeholder text I was using, which as a result will cause blank space on the page. Will revert back if needed though. Have also added a temporary border to it so we have a visual on where exactly the question text area is