Now loads test data from studentQuizLink table
The test data now loads from the studentQuizLink table. It currently loads every test in the table rather than ones assigned to a specific student as I'm not able to log in yet, but will correct as soon as possible. Still some more things to do though, such as the score being loaded, and having an actual test completion date, rather than just using the date the test was assigned (only done as a placeholder)