Fixed a bug to do with the user's answer choice
The user's answer is supposed to be 1,2,3 or 4, depending on which answer they click which it correctly does for question 1, but for question 2 it sets their answer to either 5,6,7 or 8 which it shouldn't do. This was fixed by changing the answer selection to be within the question container. Still need to make it so it only sets the user's choice after clicking the next question button.