Have made requested changes
Added checks to see if the user is a student and if they aren't then they will be given an error message. Have added a check to see if the user is logged in and if they aren't then they will be redirected to the login page. Have also replaced the SQL queries with procedures