Chart now displays scores correctly
The chart originally displayed the quiz names and their corresponding percentages in the opposite order, the percentages were being displayed correctly as newest to oldest but the quiz names were being displayed as oldest to newest. This was caused by them being called separately, and was fixed by modifying the GetStudentsQuizPercentages procedure to also get the names of each quiz and making it so it only retrieves the quizzes where the completed field = 1 as beforehand the chart was also displaying quizzes that hadn't been completed.