Fixed bug to do with fetching questions of each quiz
There was an issue to do with fetching the questions for each quiz. This was noticed after creating a maths quiz with 10 questions and it would display a different amount on the testing page (for example only 4). The bug was caused by having the limit parameter set to the QuizID rather than the limit variable